refactor: int64_t -> int8_t

+++ include/sinsei_umiusi_control/controller/thruster_controller.hpp
@@ -36,7 +36,7 @@ class ThrusterController : public controller_interface::ChainableControllerInter
     suc::state::thruster::Rpm rpm;
 
     // Thruster ID (1~4)
-    int64_t id;
+    int8_t id;
 
   public:
     ThrusterController() = default;
